+++ src/main/java/um/tesoreria/report/hexagonal/chequeras/application/service/ChequerasReportService.java
@@ -3,6 +3,7 @@ package um.tesoreria.report.hexagonal.chequeras.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import um.tesoreria.report.hexagonal.chequeras.domain.ports.in.GeneratePlanillaDetalleUseCase;
+import um.tesoreria.report.hexagonal.chequeras.domain.ports.in.GeneratePlanillaDetalleVerticalUseCase;
 import um.tesoreria.report.hexagonal.chequeras.domain.ports.in.GeneratePlanillaPagosUseCase;
 
 @Service
@@ -10,10 +11,15 @@ import um.tesoreria.report.hexagonal.chequeras.domain.ports.in.GeneratePlanillaP
 public class ChequerasReportService {
 
     private final GeneratePlanillaDetalleUseCase generatePlanillaDetalleUseCase;
+    private final GeneratePlanillaDetalleVerticalUseCase generatePlanillaDetalleVerticalUseCase;
     private final GeneratePlanillaPagosUseCase generatePlanillaPagosUseCase;
 
     public String generatePlanillaDetalle(Integer facultadId, Integer lectivoId, Integer geograficaId) {
         return generatePlanillaDetalleUseCase.generatePlanillaDetalle(facultadId, lectivoId, geograficaId);
+    }
+
+    public String generatePlanillaDetalleVertical(Integer facultadId, Integer lectivoId, Integer geograficaId) {
+        return generatePlanillaDetalleVerticalUseCase.generatePlanillaDetalle(facultadId, lectivoId, geograficaId);
     }
 
     public String generatePlanillaPagos(Integer facultadId, Integer tipoChequeraId, Integer lectivoId) {